Ok, this seems to be a winning combination. I get lock solid 240ns pulse no matter what I do.
We are writing repeatedly to the SET register to make the delay/

We are setting the PRU interconnect priority to 3 with a kernal driver.

We are not doing anything else. No T1H phase, no reading pixels from memory.

We will see if this holds up and if the kernel driver is even nessisary.